+++ OBBEdit/unpack.py
@@ -295,7 +295,7 @@ def rsg_extract(RSG_NAME, file, pathout_data, out, pathout, level):
 									open(file_path, "wb").write(file_data)
 									print("wrote " + relpath(file_path, pathout))
 								except Exception as e:
-									error_message(type(e).__name__ + " in " + file.name + ": " + RSG_NAME + ":" + DECODED_NAME + "pos: " + source.tell()-1 + ": " + str(e))
+									error_message(e, " in " + file.name + ": " + RSG_NAME + ":" + DECODED_NAME + "pos: " + source.tell())
 							# elif IS_IMAGE:
 							# 	try:
 							# 	file_path = osjoin(out, splitext(DECODED_NAME)[0] + ".PNG")
@@ -319,7 +319,7 @@ def rsg_extract(RSG_NAME, file, pathout_data, out, pathout, level):
 							print("wrote " + relpath(file_path, pathout))
 				temp = file.tell()
 	except Exception as e:
-		error_message(type(e).__name__ + " while extracting " + file.name + str(e))
+		error_message(e, " while extracting " + file.name)
 
 #def rsb_extract(file, out, level, image_decoders, pathout):
 def rsb_extract(file, pathout_data, out, level, pathout):
@@ -449,7 +449,7 @@ def file_to_folder(inp, out, level, extensions, pathout):
 			else:
 				warning_message("UNKNOWN HEADER " + HEADER.hex() + inp)
 		except Exception as e:
-			error_message("Failed OBBUnpack: " + type(e).__name__ + " in " + inp + " pos " + str(file.tell()) + ": " + str(e))
+			error_message(e, " in " + inp + " pos " + str(file.tell()), "Failed OBBUnpack: ")
 	elif isdir(inp):
 		makedirs(out, exist_ok = True)
 		for entry in sorted(listdir(inp)):
@@ -478,7 +478,7 @@ def conversion(inp, out, level, extensions, noextensions, pathout):
 			elif check[-5:] != ".json":
 				warning_message("No RTON " + inp)
 		except Exception as e:
-			error_message(type(e).__name__ + " in " + inp + " pos " + str(file.tell() -1) + ": " + str(e))
+			error_message(e, " in " + inp + " pos " + str(file.tell()))
 	elif isdir(inp):
 		makedirs(out, exist_ok = True)
 		for entry in listdir(inp):
@@ -504,10 +504,10 @@ try:
 	logerror.check_version(3, 9, 0)
 
 	print("""\033[95m
-\033[1mOBBUnpacker v1.1.7e (c) 2022 Nineteendo\033[22m
+\033[1mOBBUnpacker v1.1.7f (c) 2022 Nineteendo\033[22m
 \033[1mCode based on:\033[22m Luigi Auriemma, Small Pea & 1Zulu
 \033[1mDocumentation:\033[22m Watto Studios, YingFengTingYu, TwinKleS-C & h3x4n1um
-\033[1mFollow PyVZ2 development:\033[22m \033[4mdiscord.gg/CVZdcGKVSw\033[24m
+\033[1mFollow PyVZ2 development:\033[22m \033[4mhttps://discord.gg/CVZdcGKVSw\033[24m
 \033[0m""")
 	options = logerror.load_template(options, osjoin(application_path, "options"), 1)
 	level_to_name = ["SPECIFY", "SMF", "RSB", "RSG", "SECTION", "ENCRYPTED", "ENCODED", "DECODED"]
@@ -600,7 +600,7 @@ try:
 
 	logerror.finish_program("finished patching in", start_time)
 except Exception as e:
-	error_message(type(e).__name__ + " : " + str(e))
+	error_message(e)
 except BaseException as e:
 	warning_message(type(e).__name__ + " : " + str(e))
 logerror.close() # Close log